+++ presentation/src/main/java/com/madrid/presentation/component/movioCards/RateIcon.kt
@@ -1,4 +1,4 @@
-package com.example.presentation.component.movioCards
+package com.madrid.presentation.component.movioCards
 
 
 import androidx.compose.foundation.layout.Arrangement
