+++ presentation/src/main/java/com/madrid/presentation/component/screens/SearchScreen/SearchUiState.kt
@@ -1,4 +1,4 @@
-package com.madrid.presentation.component.screens.SearchScreen
+package com.madrid.presentation.component.screens.searchScreen
